# swagger.yaml
openapi: 3.0.0
info:
  title: FEMCS API
  description: Fire Extinguisher Management and Compliance System
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local API Gateway
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /extinguishers:
    get:
      tags: [Extinguishers]
      summary: List extinguishers
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [active, expired, serviced, decommissioned, pending_inspection] }
        - in: query
          name: type
          schema: { type: string, enum: [CO2, Dry Powder, Foam, Water] }
        - in: query
          name: compliance
          schema: { type: string, enum: [compliant, warning, critical, non_compliant] }
      responses:
        '200': { description: List of extinguishers }
    post:
      tags: [Extinguishers]
      summary: Register extinguisher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serialNumber, type, capacityLiters, manufactureDate, purchaseDate, expiryDate, location, customerId]
              properties:
                serialNumber: { type: string }
                type: { type: string, enum: [CO2, Dry Powder, Foam, Water] }
                capacityLiters: { type: number }
                manufactureDate: { type: string, format: date }
                purchaseDate: { type: string, format: date }
                expiryDate: { type: string, format: date }
                location: { type: string }
                customerId: { type: string, format: uuid }
      responses:
        '201': { description: Extinguisher registered }
...
  /extinguishers/{id}:
    get:
      tags: [Extinguishers]
      summary: Get extinguisher
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Extinguisher data }
    put:
      tags: [Extinguishers]
      summary: Update extinguisher
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200': { description: Updated }
    delete:
      tags: [Extinguishers]
      summary: Remove extinguisher (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Removed }
...